+++ src/components/Auth/Login/LoginSelector.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import { configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import {Link} from 'react-router-dom'
+import { Link } from 'react-router-dom';
 import CompanyOrCandidate from '../components/CompanyOrCandidate';
 
 configure({ adapter: new Adapter() });
@@ -13,14 +13,14 @@ test('renders without crashing', () => {
 
 test('renders a Link to "/companysignup" ', () => {
   const wrapper = shallow(<CompanyOrCandidate />);
-  const companySignup = <Link to="/companysignup"> Employer </Link>
+  const companySignup = <Link to="/companysignup"> Employer </Link>;
 
   expect(wrapper.contains(companySignup)).toEqual(true);
 });
 
 test('renders a Link to "/candidatesignup" ', () => {
   const wrapper = shallow(<CompanyOrCandidate />);
-  const candidateSignup = <Link to="/candidatesignup"> Candidate </Link>
+  const candidateSignup = <Link to="/candidatesignup"> Candidate </Link>;
 
   expect(wrapper.contains(candidateSignup)).toEqual(true);
 });
