+++ src/index.js
@@ -2,9 +2,20 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
-import * as serviceWorker from './serviceWorker';
 import { Router } from 'react-router-dom';
 import history from './history';
+import Amplify from 'aws-amplify';
+import config from './config.json';
+import * as serviceWorker from './serviceWorker';
+
+Amplify.configure({
+  Auth: {
+    mandatorySignIn: true,
+    region: config.cognito.REGION,
+    userPoolId: config.cognito.CANDIDATE_POOL_ID,
+    userPoolWebClientId: config.cognito.APP_CLIENT_ID,
+  },
+});
 
 ReactDOM.render(
   <Router history={history}>
