+++ src/router.js
@@ -1,19 +1,27 @@
-import React, { Component } from 'react';
+import React, { Component, Fragment } from 'react';
 import { Switch, Route } from 'react-router-dom';
 
 //components
-import CompanyForm from './components/forms/CompanyForm';
-import CandidateForm from './components/forms/CandidateForm';
 import LandingPage from './components/LandingPage';
-import CompanyOrCandidate from './components/CompanyOrCandidate';
+//Auth
+//Signup
+import SignupSelector from './components/Auth/Signup/SignUpSelector';
+import CompanySignup from './components/Auth/Signup/CompanySignup/CompanySignup';
+import CandidateSignup from './components/Auth/Signup/CandidateSignup/CandidateSignup';
+//Login
+import LoginSelector from './components//Auth/Login/LoginSelector/LoginSelector';
+import UserLogin from './components/Auth/Login/UserLogin/UserLogin';
+import CandidateLogin from './components/Auth/Login/CandidateLogin/CandidateLogin';
+
+//CandidatePages
 import Candidates from './components/candidate/Candidate';
 import Companies from './components/company/Companies';
-import UserLogin from './components/Login/UserLogin/UserLogin';
-import CandidateLogin from './components/Login/CandidateLogin/CandidateLogin';
-import LoginSelector from './components/Login/LoginSelector/LoginSelector';
+
+//UserCompanyPages
+
 import CompanyPositions from './components/company/CompanyPositions';
 import Positions from './components/company/compostTest';
-import NewPositionForm from './components/forms/NewPositionForm';
+import NewPositionForm from './components/NewPositionSignup/NewPositionForm';
 import CandidateAccountView from './components/candidate/CandidateAccountView';
 import TestComp from './components/company/testComp';
 
@@ -25,34 +33,35 @@ class Router extends Component {
       isLoggedInCandidate: false,
     };
   }
-  componentDidMount() {}
 
   render() {
     return (
       <Switch>
         <Route exact path="/" component={LandingPage} />
-        <Route path="/signup" component={CompanyOrCandidate} />
+        <Route path="/signup" component={SignupSelector} />
         <Route path="/login" component={LoginSelector} />
-        <Route path="/companysignup" component={CompanyForm} />
-        <Route path="/candidatesignup" component={CandidateForm} />
+        <Route path="/companysignup" component={CompanySignup} />
+        <Route path="/candidatesignup" component={CandidateSignup} />
         <Route path="/userlogin" component={UserLogin} />
         <Route path="/candidatelogin" component={CandidateLogin} />
         {this.state.isLoggedInUser && (
-          <Route path="/candidates" component={Candidates} />
+          <Fragment>
+            <Route path="/candidates" component={Candidates} />
+            <Route path="/newposition" component={NewPositionForm} />
+            <Route path="/positions" component={CompanyPositions} />
+          </Fragment>
         )}
         {this.state.isLoggedInCandidate && (
-          <Route path="/companies" component={Companies} />
+          <Fragment>
+            <Route
+              path="/candidateAccount/:id"
+              component={CandidateAccountView}
+            />
+            <Route path="/positions" component={CompanyPositions} />
+            <Route path="/companies" component={Companies} />
+            <Route exact path="/companies" component={Companies} />
+          </Fragment>
         )}
-        <Route path="/companysignup" component={CompanyForm} />
-        <Route path="/candidatesignup" component={CandidateForm} />
-        <Route path="/candidates" component={Candidates} />
-        <Route path="/newposition" component={NewPositionForm} />
-        <Route path="/positions" component={CompanyPositions} />
-        <Route exact path="/companies" component={Companies} />
-        <Route path="/candidateAccount/:id" component={CandidateAccountView} />
-        <Route path="/login" component={LoginSelector} />
-        <Route path="/candidatelogin" component={CandidateLogin} />
-        <Route path="/userlogin" component={UserLogin} />
 
         <Route exact path="/companies/:id" component={TestComp} />
